aply changes of wordmatch to avalonia and oxford

+++ diagrammatically.Domein/WordMatch.cs
@@ -4,11 +4,11 @@
     {
         public string Search { get; }
         public string Word { get; }
-        public int Match { get; }
+        public double Match { get; }
         public int Used { get; }
         public string Sourse { get; }
 
-        public WordMatch(string search, string word, int match, int used, string sourse)
+        public WordMatch(string search, string word, double match, int used, string sourse)
         {
             Search = search;
             Word = word;
